feat(ui): promote MCP servers to first-class field in NewSessionModal
Mirrors the SessionSettingsModal layout — MCP servers picker is now
visible in the primary zone of the create-session form, instead of
buried inside the "Agentic Tool Configuration" collapse. The long-tail
agent settings (model / permission mode / effort / codex) stay
collapsed.

- Extract the Form.Item + service-readable gate into a small reusable
  SessionMcpServersField component (drop-in for both modals).
- Add hideMcpServers prop to AgenticToolConfigForm so callers that
  render MCP elsewhere can suppress the bundled field. Other 6 callers
  (ForkSpawn / Schedule / DefaultAgenticSettings / GatewayChannels /
  UserSettings / ZoneTrigger) keep current behavior via the default.
