+++ apps/agor-ui/src/components/MCPServerSelect/index.ts
@@ -1,1 +1,3 @@
 export { MCPServerSelect } from './MCPServerSelect';
+export type { SessionMcpServersFieldProps } from './SessionMcpServersField';
+export { SessionMcpServersField } from './SessionMcpServersField';
